11/23/22
ZyBooks keeps giving me an EOF Error with this code but when I put the exact same code in IDLE, I don't get the EOF and it works as intended. Not exactly sure what the issue is, but if I do it without the while loop, ZyBooks won't even allow any more years after the first test to take. I figured I'd go ahead and upload this since it works just fine in IDLE.